// src/types/style.types.ts
export type Flex = {
  justify?: string;
  align?: string;
  row?: string;
  height?: string;
  margin?: string;
  width?: string;
};

export type Span = {
  color?: string;
  deco?: string;
  size?: string;
  margin?: string;
};
export type AbsoluteType = {
  top?: string;
  left?: string;
  right?: string;
};